FU-106 refactor: group form parts
