+++ src/containers/customer/reservation/submit/parts/select-option-form.tsx
@@ -1,6 +1,6 @@
 import Dropdown from "@/components/common/dropdown";
 import { Option } from "product-types";
-import PartLayout from "./part-layout";
+import PartLayout from "../part-layout";
 
 const SelectOptionForm = ({ options }: { options: Option[] }) => {
   return (
